refactor: move CreateAppointmentScreen logic to its own file
